backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
google-generativeai==0.3.2
requests==2.31.0
pydantic==2.6.1
firecrawl==0.1.0
python-multipart==0.0.9 